Moving hardcoded tests into a separate regression test file
